scripts/requirements.txt
supabase==2.10.0
faker==33.1.0
python-dotenv==1.0.1
